pyproject.toml
[project]
name = "yolo-toolchain"
version = "1.0.0"
description = "Enhanced YOLOv11 toolchain for object detection"
readable = "README.md"
requires-python = ">=3.8"
license = {text = "AGPL-3.0"}
authors = [
    {name = "Matrix Agent"}
]
keywords = ["yolo", "object-detection", "deep-learning", "computer-vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "ultralytics>=8.0.0",
    "numpy>=1.21.0",
    "opencv-python>=4.5.0",
    "Pillow>=9.0.0",
    "albumentations>=1.3.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "pyyaml>=6.0",
    "tqdm>=4.62.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "huggingface_hub>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
]
# macOS only - CoreML support
coreml = ["coremltools"]
# Monitoring
wandb = ["wandb"]
tensorboard = ["tensorboard"]
# GPU acceleration (install manually on Linux/Windows with NVIDIA GPU)
# pip install tensorrt --extra-index-url https://pypi.nvidia.com/
# Intel CPU optimization (install manually on Linux)
# pip install openvino-dev

[project.scripts]
yolo-convert = "src.tools.dataset_builder:main"
yolo-augment = "src.tools.augmentor:main"
yolo-auto-annotate = "src.tools.auto_annotator:main"
yolo-verify = "src.tools.verify_annotator:main"
yolo-preprocess = "src.tools.preprocess:main"
yolo-download = "src.tools.downloader:main"
yolo-train = "src.train.trainer:main"
yolo-freeze-train = "src.train.freeze_trainer:main"
yolo-incremental-train = "src.train.incremental_trainer:main"
yolo-diagnose = "src.eval.diagnostics:main"
yolo-export = "src.export.exporter:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
